+++ src/cctag/Fitting.hpp
@@ -16,7 +16,7 @@ namespace numerical {
 
 // Precondition : pts.size >=2
 // TODO déplacer innerProdMin
-double innerProdMin( const std::vector<cctag::EdgePoint*>& childrens, double thrCosDiffMax, Point2d<Eigen::Vector3i> & p1, Point2d<Eigen::Vector3i> & p2 );
+float innerProdMin( const std::vector<cctag::EdgePoint*>& childrens, float thrCosDiffMax, Point2d<Eigen::Vector3i> & p1, Point2d<Eigen::Vector3i> & p2 );
 
 void circleFitting(cctag::numerical::geometry::Ellipse& e, const std::vector<cctag::EdgePoint*>& points);
 
@@ -26,9 +26,9 @@ void ellipseFitting( cctag::numerical::geometry::Ellipse& e, const std::vector<c
 
 void ellipseFitting( cctag::numerical::geometry::Ellipse& e, const std::list<cctag::EdgePoint*>& childrens );
 
-void ellipse_fit_with_gradients( double *pts, double *grad, int pts_size,
-                                 double **buff, int *size_buff_max,
-                                 double *param );
+void ellipse_fit_with_gradients( float *pts, float *grad, int pts_size,
+                                 float **buff, int *size_buff_max,
+                                 float *param );
 
 void ellipseFittingWithGradientsToto( const std::vector<EdgePoint *> & vPoint, cctag::numerical::geometry::Ellipse & ellipse );
