+++ src/cctag/Level.cpp
@@ -47,8 +47,8 @@ Level::~Level( )
 }
 
 void Level::setLevel( const cv::Mat & src,
-                      const double thrLowCanny,
-                      const double thrHighCanny,
+                      const float thrLowCanny,
+                      const float thrHighCanny,
                       const cctag::Parameters* params )
 {
     if( _cuda_allocates ) {
